Added module 'attempt'
